LibWeb: Split BackgroundStyleValue out of StyleValue.{h,cpp}
