+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -502,62 +502,6 @@ struct StyleValueWithDefaultOperators : public StyleValue {
     }
 };
 
-class BackgroundStyleValue final : public StyleValueWithDefaultOperators<BackgroundStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<BackgroundStyleValue> create(
-        ValueComparingNonnullRefPtr<StyleValue const> color,
-        ValueComparingNonnullRefPtr<StyleValue const> image,
-        ValueComparingNonnullRefPtr<StyleValue const> position,
-        ValueComparingNonnullRefPtr<StyleValue const> size,
-        ValueComparingNonnullRefPtr<StyleValue const> repeat,
-        ValueComparingNonnullRefPtr<StyleValue const> attachment,
-        ValueComparingNonnullRefPtr<StyleValue const> origin,
-        ValueComparingNonnullRefPtr<StyleValue const> clip)
-    {
-        return adopt_ref(*new BackgroundStyleValue(move(color), move(image), move(position), move(size), move(repeat), move(attachment), move(origin), move(clip)));
-    }
-    virtual ~BackgroundStyleValue() override = default;
-
-    size_t layer_count() const { return m_properties.layer_count; }
-
-    auto attachment() const { return m_properties.attachment; }
-    auto clip() const { return m_properties.clip; }
-    auto color() const { return m_properties.color; }
-    auto image() const { return m_properties.image; }
-    auto origin() const { return m_properties.origin; }
-    auto position() const { return m_properties.position; }
-    auto repeat() const { return m_properties.repeat; }
-    auto size() const { return m_properties.size; }
-
-    virtual ErrorOr<String> to_string() const override;
-
-    bool properties_equal(BackgroundStyleValue const& other) const { return m_properties == other.m_properties; }
-
-private:
-    BackgroundStyleValue(
-        ValueComparingNonnullRefPtr<StyleValue const> color,
-        ValueComparingNonnullRefPtr<StyleValue const> image,
-        ValueComparingNonnullRefPtr<StyleValue const> position,
-        ValueComparingNonnullRefPtr<StyleValue const> size,
-        ValueComparingNonnullRefPtr<StyleValue const> repeat,
-        ValueComparingNonnullRefPtr<StyleValue const> attachment,
-        ValueComparingNonnullRefPtr<StyleValue const> origin,
-        ValueComparingNonnullRefPtr<StyleValue const> clip);
-
-    struct Properties {
-        ValueComparingNonnullRefPtr<StyleValue const> color;
-        ValueComparingNonnullRefPtr<StyleValue const> image;
-        ValueComparingNonnullRefPtr<StyleValue const> position;
-        ValueComparingNonnullRefPtr<StyleValue const> size;
-        ValueComparingNonnullRefPtr<StyleValue const> repeat;
-        ValueComparingNonnullRefPtr<StyleValue const> attachment;
-        ValueComparingNonnullRefPtr<StyleValue const> origin;
-        ValueComparingNonnullRefPtr<StyleValue const> clip;
-        size_t layer_count;
-        bool operator==(Properties const&) const = default;
-    } m_properties;
-};
-
 class BackgroundRepeatStyleValue final : public StyleValueWithDefaultOperators<BackgroundRepeatStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<BackgroundRepeatStyleValue> create(Repeat repeat_x, Repeat repeat_y)
